+++ 1 Presentation layer/CandidateEvaluator.Server/appsettings.json
@@ -3,6 +3,7 @@
     "ConnectionString": "UseDevelopmentStorage=true;",
     "CategoryTableName": "CECategory",
     "QuestionTableName": "CEQuestion",
+    "InterviewTableName": "CEInterview",
     "UserRecentActivityTableName": "CERecentActivity"
   },
   "AadOptions": {
