+++ 2 Business layer/CandidateEvaluator.Core/Handlers/Commands/Category/CreateCategoryHandler.cs
@@ -1,15 +1,11 @@
-﻿using CandidateEvaluator.Contract.Commands.Category;
+﻿using System;
+using System.Threading.Tasks;
+using CandidateEvaluator.Contract.Commands.Category;
 using CandidateEvaluator.Contract.Handlers;
 using CandidateEvaluator.Contract.Models;
 using CandidateEvaluator.Contract.Repositories;
-using CandidateEvaluator.Contract.Services;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace CandidateEvaluator.Core.Handlers.Commands
+namespace CandidateEvaluator.Core.Handlers.Commands.Category
 {
     public class CreateCategoryHandler : ICommandHandler<CreateCategory>
     {
@@ -25,7 +21,7 @@ namespace CandidateEvaluator.Core.Handlers.Commands
 
         public async Task<Guid> HandleAsync(CreateCategory command)
         {
-            var model = new Category
+            var model = new Contract.Models.Category
             {
                 OwnerId = command.OwnerId,
                 Name = command.Name
