+++ 2 Business layer/CandidateEvaluator.Core/Handlers/Commands/Category/UpdateCategoryHandler.cs
@@ -1,12 +1,11 @@
-﻿using CandidateEvaluator.Contract.Commands.Category;
+﻿using System;
+using System.Threading.Tasks;
+using CandidateEvaluator.Contract.Commands.Category;
 using CandidateEvaluator.Contract.Handlers;
 using CandidateEvaluator.Contract.Models;
 using CandidateEvaluator.Contract.Repositories;
-using CandidateEvaluator.Contract.Services;
-using System;
-using System.Threading.Tasks;
 
-namespace CandidateEvaluator.Core.Handlers.Commands
+namespace CandidateEvaluator.Core.Handlers.Commands.Category
 {
     public class UpdateCategoryHandler : ICommandHandler<UpdateCategory>
     {
@@ -22,14 +21,14 @@ namespace CandidateEvaluator.Core.Handlers.Commands
 
         public async Task<Guid> HandleAsync(UpdateCategory command)
         {
-            var model = new Category
+            var model = new Contract.Models.Category
             {
                 Id = command.Id,
                 OwnerId = command.OwnerId,
                 Name = command.Name
             };
-
             await _modelRepository.Update(model);
+
             await _activityRepository.Upsert(model.OwnerId, new RecentActivity
             {
                 Type = EntityType.Category,
