+++ 2 Business layer/CandidateEvaluator.Core/Handlers/Queries/Question/GetQuestionHandler.cs
@@ -1,12 +1,11 @@
-﻿using CandidateEvaluator.Contract.Handlers;
-using CandidateEvaluator.Contract.Models;
+﻿using System.Threading.Tasks;
+using CandidateEvaluator.Contract.Handlers;
 using CandidateEvaluator.Contract.Queries.Question;
 using CandidateEvaluator.Contract.Repositories;
-using System.Threading.Tasks;
 
-namespace CandidateEvaluator.Core.Handlers.Queries
+namespace CandidateEvaluator.Core.Handlers.Queries.Question
 {
-    public class GetQuestionHandler : IQueryHandler<GetQuestion, Question>
+    public class GetQuestionHandler : IQueryHandler<GetQuestion, Contract.Models.Question>
     {
         private readonly IQuestionRepository _modelRepository;
 
@@ -15,7 +14,7 @@ namespace CandidateEvaluator.Core.Handlers.Queries
             _modelRepository = modelRepository;
         }
 
-        public Task<Question> HandleAsync(GetQuestion query)
+        public Task<Contract.Models.Question> HandleAsync(GetQuestion query)
         {
             return _modelRepository.Get(query.OwnerId, query.CategoryId, query.Id);
         }
